pyproject.toml
[tool.poetry]
name = "repl_nix_GifMOD"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.2.2"
Flask = "^2.2.3"
python-dotenv = "^1.0.0"
gunicorn = "20.1.0"
waitress = "^2.1.2"
schedule = "^1.1.0"
botlib = "^0.0.0"
APScheduler = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
